+++ crates/cervo-core/src/batcher/scratch.rs
@@ -175,6 +175,10 @@ impl ScratchPad {
     pub(crate) fn output_name(&self, slot: usize) -> &str {
         &self.outputs[slot].name
     }
+
+    pub(crate) fn lookup_output_slot(&self, name: &str) -> Option<usize> {
+        self.outputs.iter().position(|slot| slot.name == name)
+    }
 }
 
 /// A view over a set of batch elements in a scratch pad.
@@ -184,6 +188,34 @@ pub struct ScratchPadView<'a> {
 }
 
 impl<'a> ScratchPadView<'a> {
+    pub fn inner(&self) -> &ScratchPad {
+        self.pad
+    }
+
+    /// View of the input at location `slot`.
+    pub fn input_slot_with_id(&self, slot: usize) -> (&[u64], &[f32]) {
+        (
+            &self.pad.ids[self.batch_range.clone()],
+            self.pad.input_slot(slot, self.batch_range.clone()),
+        )
+    }
+
+    /// Mutable view of the input at location `slot`.
+    pub fn input_slot_mut_with_id(&mut self, slot: usize) -> (&[u64], &mut [f32]) {
+        (
+            &self.pad.ids[self.batch_range.clone()],
+            self.pad.inputs[slot].view_mut(self.batch_range.clone()),
+        )
+    }
+
+    /// Mutable view of the input at location `slot`.
+    pub fn output_slot_mut_with_id(&mut self, slot: usize) -> (&[u64], &mut [f32]) {
+        (
+            &self.pad.ids[self.batch_range.clone()],
+            self.pad.outputs[slot].view_mut(self.batch_range.clone()),
+        )
+    }
+
     /// View of the input at location `slot`.
     pub fn input_slot(&self, slot: usize) -> &[f32] {
         self.pad.input_slot(slot, self.batch_range.clone())