+++ crates/cervo-core/src/epsilon.rs
@@ -123,6 +123,8 @@ pub struct EpsilonInjector<T: Inferer, NG: NoiseGenerator = HighQualityNoiseGene
     count: usize,
     index: usize,
     generator: NG,
+
+    inputs: Vec<(String, Vec<usize>)>,
 }
 
 impl<T> EpsilonInjector<T, HighQualityNoiseGenerator>
@@ -159,11 +161,19 @@ where
             None => bail!("model has no input key {:?}", key),
         };
 
+        let inputs = inputs
+            .iter()
+            .filter(|(k, _)| *k != key)
+            .map(|(k, v)| (k.to_owned(), v.to_owned()))
+            .collect::<Vec<_>>();
+
         Ok(Self {
             inner: inferer,
             index,
             count,
             generator,
+
+            inputs,
         })
     }
 }
@@ -177,7 +187,7 @@ where
         self.inner.select_batch_size(max_count)
     }
 
-    fn infer_raw(&self, mut batch: ScratchPadView<'_>) -> Result<(), anyhow::Error> {
+    fn infer_raw(&self, batch: &mut ScratchPadView<'_>) -> Result<(), anyhow::Error> {
         let total_count = self.count * batch.len();
         let output = batch.input_slot_mut(self.index);
         self.generator.generate(total_count, output);
@@ -186,10 +196,22 @@ where
     }
 
     fn input_shapes(&self) -> &[(String, Vec<usize>)] {
-        self.inner.input_shapes()
+        &self.inputs
     }
 
-    fn output_shapes(&self) -> &[(String, Vec<usize>)] {
-        self.inner.output_shapes()
+    fn raw_input_shapes(&self) -> &[(String, Vec<usize>)] {
+        self.inner.raw_input_shapes()
+    }
+
+    fn raw_output_shapes(&self) -> &[(String, Vec<usize>)] {
+        self.inner.raw_output_shapes()
+    }
+
+    fn begin_agent(&mut self, id: u64) {
+        self.inner.begin_agent(id);
+    }
+
+    fn end_agent(&mut self, id: u64) {
+        self.inner.end_agent(id);
     }
 }